+++ src/classes/Employee.java
@@ -3,7 +3,7 @@ package classes;
 public class Employee extends People{
     private float salary;
 
-    public Employee(String name, int age, int height, int weight, String cpf, float salary) {
+    public Employee(String name, int age, int height, float weight, String cpf, float salary) {
         super(name, age, height, weight, cpf);
         this.salary = salary;
     }
