+++ src/classes/Veterinary.java
@@ -3,7 +3,7 @@ package classes;
 public class Veterinary extends Employee{
     private int crmv;
 
-    public Veterinary(String name, int age, int height, int weight, String cpf, float salary, int crmv) {
+    public Veterinary(String name, int age, int height, float weight, String cpf, float salary, int crmv) {
         super(name, age, height, weight, cpf, salary);
         this.crmv = crmv;
     }
